Propagate StoreRemoved, ObjectAdded and ObjectRemoved events.
These events, along with the previously exposed StoreAdded event,
are the most frequently implemented events on MTP devices. In the
case of Google's Android MTP stack, they are the only events
exposed.

In the case of StoreRemoved, the param value is the storage ID
and in the case of ObjectAdded/Removed, it is the entity ID.

Signed-off-by: Philip Langdale <philipl@overt.org>
Signed-off-by: Linus Walleij <triad@df.lth.se>