PATCH: avoid unconditional clear_halt
Hi,

During debugging weird "access only possible once" scenarios,
on Raspberry Pi and also on regular Linux with USB 3.0 controllers,
I found that clearing halt on a endpoint which is not stalled
will confuse the (kernel) driver and device and make them really
stall until resetted.

This is probably a Kernel bug.

So do not do a unconditional clear_halt.

I am leaving the conditional clear_stall in, as it queries the
status before.

This might need testing.

Ciao, Marcus

Signed-off-by: Marcus Meissner <marcus@jet.franken.de>
Signed-off-by: Linus Walleij <triad@df.lth.se>