util: print the version of libmtp when warning about unknown device
Many reports about unknown devices are done copy-pasting the error
string where libmtp informs about its IDs, but many reporters forget
to post the libmtp version they are using (example: [1]).

If we add the libmtp version to the same string where we report the
IDs, we might save some back&forth between the reporter and the devs
in the future.

[1] http://sourceforge.net/p/libmtp/bugs/1361/

Reviewed-by: Marcus Meissner <meissner@suse.de>