autogen.sh: call ./configure at the end
Most open source projects have the convention of
just running 3 commands to install in a system:
configuration, build, installation.

The first of these steps usually consists of running
./configure (if it's a tarball) *OR* autogen.sh (if
it's a git checkout) because autogen.sh normally
calls ./configure (if NOCONFIGURE is not set).

So then in general it is expected that calling
./autogen.sh --prefix=/foo in a git checkout
has the same effect as calling
./configure --prefix=/foo in a tarball, which
was not the case before this.