Add LIBMTP_FILES_AND_FOLDERS_ROOT and fix examples
Examples assumed that parent id 0 will list only the top directory, but
actually that's an id for all files (it does happen to correspond to the
reported parent id however). The actual id to be used for the top folder
is 0xffffffff.

Add a new define to make the value more obvious as well.