Add a workaround for the GetPartialObject offset bug on Samsung Galaxy
The MTP stack of Samsung Galaxy devices has a mysterious bug in
GetPartialObject. When GetPartialObject is invoked to read the last
bytes of a file and the amount of data to read is such that the last
USB packet sent in the reply matches exactly the USB 2.0 packet size,
then the Samsung Galaxy device hangs, resulting in a timeout error.

As a workaround, we read one less byte instead of reaching the end of
the file, forcing the caller to perform an additional read to get the
last byte (i.e. the final read that would fail is replaced with two
partial reads that succeed).

This patch fixes GitHub issue #8.

Signed-off-by: Christophe Vu-Brugier <cvubrugier@fastmail.fm>