mtp-hotplug: make use of the hwdb
There were two equivalent mechanisms implemented: hwdb entries and the
traditional udev rules. This removes the per-entry udev rules in
favour of looking up properties in the hwdb and acting based on that.
There should be no functional change.

Apart from removing duplication, the advantage is efficiency: the hwdb
uses an quick trie lookup, but it is also done by the generic udev
rules anyway, so it is essentially free. OTOH, udev rules are
processed sequentially and cannot be optimized. Udev upstream doesn't
want every project to install long list of udev rules for device
matching because this simply doesn't scale with the thousands upon
thousands of different devices that we need to provide properties for.