Properly use dirname()/basename() in connect example
Citing man 3 dirname:

On Linux both dirname() and basename() may modify the contents of path,
so it may be desirable to pass a copy when calling one of these functions.
These  functions  may  return pointers to statically allocated memory which
may be overwritten by subsequent calls. Alternatively, they may return a
pointer to some part of path, so that the string referred to by path should
not be modified or freed until the pointer returned by the function is no
longer required.

Marcus: small fix added