Add Unicode >= 2.0 extended chars to ucs2_strlen()
ucs2_strlen() was only able to act as a strlen() type function but was
inaccurate for working with chars 0x10000..0x10ffff which are coded as
pairs {0xd800..0xdbff}:{0xdc00..0xdfff} which is one char but uses two
utf-16 values. String is tested as utf-16le regardless of CPU endian.

mode=0: acts like the original ucs2_strlen(), which is still useful to
find a count useful for buffers.

mode=1: is a hybrid, which will count one char for the code pairs used
to create values 0x10000..0x10ffff, and does a soft fail to count code
as separate values if the code pair isn't grouped together. This could
be useful in some situations.

mode=2: does a strict check and returns length=-1 if the code pairs is
out of sequence. This is useful to get the right char count and verify
that the extended char code pairs are in the right sequence.