utf16_to_utf8() - Remove dependence from iconv or libiconv
Convert utf16 coded characters to utf8. This also fixes the output buffer
to allow for a worst-case of 1024 x 4 utf8 chars which is possible if all
chars are from range 0x10000..0x10ffff.
This is a subset of a converter made for fontforge, based on utf8_idpb().
See function: 2013-10-06, Expanded utf8_idpb() to output up to 0x7FFFFFFF