utf8_to_utf16() - Remove dependence from iconv or libiconv
Convert utf8 coded characters to utf16. This also fixes the output buffer
to allow for a worst-case of 1024 utf16 chars in range 0x10000..0x10ffff.
This is an optimized subset of a converter made for fontforge with added
checks for bad utf-8 codes. Default behaviour for this function returns a
zero length string. Added another mode to return NULL since it's possible
to receive zero length strings that are not errors (example: empty line).