libmtp configure putting UDEV files in the wrong location #271
Found a better solution for './configure' and the "/usr/lib/udev" problem
when running 'make distcheck' and programs like 'rpmbuild -ba' that still
required values like {prefix}. The solution was to set:
UDEV=\${prefix}/../lib/udev (for prefix=/usr/local) or
UDEV=\${prefix}/lib/udev (for prefix=/usr).

Also copied UDEV to TARGET_UDEV so that we can simplify the if/then/else,
and this simplified Makefile.am so we don't need to use the UDEV_SNIPPET.
TARGET_UDEV is then overwritten with crossbuilddir if we're crossbuilding
otherwise it remains as whatever directory UDEV originally set it to.

Also added AC_MSG_NOTICE() so we can see results of using './configure'.