housekeeping cleanup util/mtp-hotplug.c, all memory now free() at exit
LIBMTP_Init() expected to be used for normal operations, but it will
initialize a number of variables that are expected to be used later.

We only want to grab a static list, and able to avoid LIBMTP_Init().
Valgrind now returns:
==55621== HEAP SUMMARY:
==55621==     in use at exit: 0 bytes in 0 blocks
==55621==   total heap usage: 7 allocs, 7 frees, 7,038 bytes allocated
==55621==
==55621== All heap blocks were freed -- no leaks are possible
==55621==
==55621== For lists of detected and suppressed errors, rerun with: -s
==55621== ERROR SUMMARY: 0 errors from 0 contexts (suppressed: 0 from 0)

Commandline test:
libtool --mode=execute valgrind --leak-check=full --show-leak-kinds=all
 util/mtp-hotplug -g"nobody" -m"666" -g"anybody" -m"555"